refactor(elaborator): guard-managed Scope for the annotate walk (S1)
Introduce elaborator/scope.rs: Scope (was trait_env::AnnotateCtx) now
also carries default_scope_module, and effect params move onto
TraitContext (they are generic-scope state, so TypeParamScope restores
them; the reader-less current_effect_params set is deleted). Every
manual scope save/restore is replaced by a panic-safe guard:

- Item::Impl arm extracted into resolve_impl_item, wrapped by
  TypeParamScope instead of the hand-rolled trait_ctx clone with three
  restore exits.
- self_type saves in method_lookup.rs go through with_self_type_override.
- default_scope_module replace/restore pairs in call.rs / expr.rs go
  through with_default_scope_module.
- The six effect-param take/install/restore pairs collapse to
  TraitContext::install_effect_params under the enclosing guard.

WEP 2026-07-10 S1; scope fields are now mutated only through scope.rs.
