+++ wado-compiler/src/elaborator/trait_env.rs
@@ -4,17 +4,14 @@
 //! It provides O(1) lookup of trait implementations by type name and trait name,
 //! replacing linear scans across all modules.
 
-use std::cell::RefCell;
-use std::ops::{Deref, DerefMut};
 use std::sync::Arc;
 
 use crate::ast::{self, AstId, Item, Module, Type};
-use crate::compiler_host::CompilerHost;
 use crate::hashmap::{IndexMap, IndexSet};
 use crate::kiln::InvocationIndex;
 use crate::module_source::{ModuleSource, ModuleSourceInterner};
 use crate::name;
-use crate::tir::{TypeId, TypeTable};
+use crate::tir::TypeTable;
 
 /// A module's type-name import scope, derived once from its `use`
 /// declarations. The single source of truth for how a module resolves a type
@@ -179,7 +176,6 @@ fn collect_case_names(
     }
 }
 
-use super::Elaborator;
 use super::types::TypeError;
 use crate::symbol::SymbolTable;
 
@@ -1329,239 +1325,6 @@ fn check_all_orphan_rules(
     }
 
     violations
-}
-
-/// Mutable trait resolution context scoped to the current resolution site.
-///
-/// Groups all state that changes when entering/leaving generic scopes
-/// (impl blocks, trait method lookups, etc). Use [`Elaborator::enter_fresh_type_param_scope`]
-/// or [`Elaborator::enter_inherited_type_param_scope`] to mutate this safely with RAII
-/// restore on drop.
-#[derive(Clone, Default)]
-pub(super) struct TraitContext {
-    /// Type parameters currently in scope (name → (index, `TypeId`)).
-    /// Set when resolving generic structs, functions, or impl blocks.
-    pub(super) type_params: IndexMap<String, (u32, TypeId)>,
-    /// `AstId` of each type param's declaration site (for LSP jump-to-def on
-    /// type-parameter uses). Parallel to `type_params`, keyed by name.
-    pub(super) type_param_decls: IndexMap<String, ast::AstId>,
-    /// Trait bounds on type parameters in scope (name → full bounds with assoc types).
-    /// Used for resolving trait methods on type params (e.g., `T.cmp()` when T: Ord).
-    pub(super) type_param_bounds: IndexMap<String, Vec<ast::TraitBound>>,
-    /// Associated type bindings in scope (`Self::Name` → resolved type).
-    /// Set when resolving trait implementations.
-    pub(super) assoc_type_bindings: IndexMap<String, TypeId>,
-    /// Current `Self` type in scope (the type being implemented in an impl block).
-    pub(super) self_type: Option<TypeId>,
-}
-
-/// Per-function annotate-time scope: the trait-resolution context
-/// ([`TraitContext`]) plus the `type_implements_trait` recursion guard,
-/// bundled so they move as one unit (queries take `&AnnotateCtx`). Neither
-/// may move onto the shared `TypeSystem`: `trait_ctx` is per-function and
-/// `trait_check_stack` is a per-call frame stack whose sharing would leak
-/// frames across module walks.
-#[derive(Default)]
-pub(super) struct AnnotateCtx {
-    pub(super) trait_ctx: TraitContext,
-    pub(super) trait_check_stack: RefCell<Vec<(TypeId, String)>>,
-}
-
-/// RAII guard that restores `Elaborator::trait_ctx` to its saved value on drop.
-///
-/// Implements `Deref<Target = Elaborator>` so it can be used as a transparent
-/// elaborator handle inside the scope. Restoration is panic-safe: even if the
-/// scope body panics, drop still runs and the parent context is reinstated.
-///
-/// Use [`Elaborator::enter_inherited_type_param_scope`] to enter a new scope.
-/// It preserves the current `trait_ctx` so the child scope can register new
-/// entries on top of the parent's. Callers that want a clean slate for a
-/// specific field (matching the legacy `mem::take` pattern) should clear that
-/// field on `scope.annotate_ctx.trait_ctx` after entering.
-pub(super) struct TypeParamScope<'r, 'a, H: CompilerHost> {
-    elaborator: &'r mut Elaborator<'a, H>,
-    saved: TraitContext,
-}
-
-impl<'a, H: CompilerHost> Deref for TypeParamScope<'_, 'a, H> {
-    type Target = Elaborator<'a, H>;
-    fn deref(&self) -> &Elaborator<'a, H> {
-        self.elaborator
-    }
-}
-
-impl<'a, H: CompilerHost> DerefMut for TypeParamScope<'_, 'a, H> {
-    fn deref_mut(&mut self) -> &mut Elaborator<'a, H> {
-        self.elaborator
-    }
-}
-
-impl<H: CompilerHost> TypeParamScope<'_, '_, H> {
-    /// Access the saved (parent) `TraitContext`. Useful when setting up an
-    /// inner scope for an impl block whose impl type refers to one of the
-    /// parent's type params (blanket impl / `&T` impl / variadic impl).
-    pub(super) fn saved(&self) -> &TraitContext {
-        &self.saved
-    }
-}
-
-impl<H: CompilerHost> Drop for TypeParamScope<'_, '_, H> {
-    fn drop(&mut self) {
-        self.elaborator.annotate_ctx.trait_ctx = std::mem::take(&mut self.saved);
-    }
-}
-
-impl<'a, H: CompilerHost> Elaborator<'a, H> {
-    /// Enter an inherited type-param scope. The current `trait_ctx` is cloned
-    /// into the saved slot, but left in place so the inner work can register
-    /// additional type params on top of what the parent already had. The
-    /// original context is restored when the returned guard is dropped.
-    ///
-    /// Callers that want a clean slate (matching the legacy
-    /// `mem::take(&mut self.annotate_ctx.trait_ctx.type_params)` pattern) should clear the
-    /// specific fields they want to reset on `scope.annotate_ctx.trait_ctx` after entering
-    /// the scope — only the fields they touch need to be cleared, all others
-    /// are inherited from the parent scope.
-    pub(super) fn enter_inherited_type_param_scope(&mut self) -> TypeParamScope<'_, 'a, H> {
-        let saved = self.annotate_ctx.trait_ctx.clone();
-        TypeParamScope {
-            elaborator: self,
-            saved,
-        }
-    }
-
-    /// Register a list of generic parameters as `TypeParam` / `TypePack` ids
-    /// in the current `trait_ctx`, starting from `offset`. Skips effect params.
-    /// Returns the next free index (i.e. `offset + non_effect_count`).
-    ///
-    /// Trait bounds attached to each parameter are also recorded in
-    /// `type_param_bounds` so trait-method lookups on the parameter work.
-    pub(super) fn register_generic_params(
-        &mut self,
-        params: &[ast::GenericParam],
-        offset: u32,
-    ) -> u32 {
-        let mut idx = offset;
-        for tp in params.iter().filter(|p| !p.is_effect) {
-            // `<F: fn(...)>` / `<F: fn mut(...)>` binds the parameter directly
-            // to the bound's function type. The closure-type bound is just
-            // surface syntax for "F is exactly this signature" — eager
-            // substitution lets `f: F` be callable inside the body and folds
-            // every callsite onto the same shared canonical closure shape.
-            //
-            // Fn-bound params do NOT consume a `TypeParam` index slot. This
-            // keeps the index space dense for real type params so the
-            // substitution map in `substitute_type_params` (which is keyed by
-            // `TypeParam.index`) lines up with the positional order used by
-            // the inference cache. Without this, mixed declarations like
-            // `<F: fn(...), T>` would leave `T` at `TypeParam(_, 1)` while
-            // the cache placed it at position 0, breaking substitution.
-            let fn_bound_sig = if tp.is_pack {
-                None
-            } else {
-                tp.bounds.iter().find_map(|b| b.fn_signature.as_ref())
-            };
-            let (type_id, consumed_index) = if tp.is_pack {
-                (
-                    self.tysys
-                        .type_table
-                        .borrow_mut()
-                        .make_type_pack(tp.name.clone(), idx),
-                    true,
-                )
-            } else if let Some(sig) = fn_bound_sig {
-                (self.resolve_type(&ast::Type::Function(sig.clone())), false)
-            } else {
-                (
-                    self.tysys
-                        .type_table
-                        .borrow_mut()
-                        .make_type_param(tp.name.clone(), idx),
-                    true,
-                )
-            };
-            self.annotate_ctx
-                .trait_ctx
-                .type_params
-                .insert(tp.name.clone(), (idx, type_id));
-            self.annotate_ctx
-                .trait_ctx
-                .type_param_decls
-                .insert(tp.name.clone(), tp.id);
-            // Filter out `fn`/`fn mut` bounds before recording (they're already
-            // realised in the bound type itself); only "real" trait bounds need
-            // remembering for method lookup.
-            let real_bounds: Vec<ast::TraitBound> = tp
-                .bounds
-                .iter()
-                .filter(|b| b.fn_signature.is_none())
-                .cloned()
-                .collect();
-            if !real_bounds.is_empty() {
-                self.annotate_ctx
-                    .trait_ctx
-                    .type_param_bounds
-                    .insert(tp.name.clone(), real_bounds);
-            }
-            if consumed_index {
-                idx += 1;
-            }
-        }
-        idx
-    }
-
-    /// Bind a trait's declared type parameters to the impl's concrete trait
-    /// arguments in the current scope. Given `trait Foo<T, U>` and an impl
-    /// instance `Foo<i32, String>`, this registers `T → i32` and `U → String`
-    /// in `trait_ctx.type_params` together with their declared bounds.
-    ///
-    /// Callers must have any impl-level type params already registered because
-    /// the trait args may reference them (e.g., `impl<X> Foo<Container<X>>`).
-    /// Trait args are resolved in the current scope before being inserted.
-    ///
-    /// Entries already present in `type_params` (e.g., re-used impl-level
-    /// names) are left untouched.
-    pub(super) fn bind_trait_type_params_from_impl(&mut self, trait_type: &ast::Type) {
-        let trait_name = self.get_type_name(trait_type);
-        let Some(trait_decl_type_params) = self.find_trait_decl_type_params(&trait_name) else {
-            return;
-        };
-        let trait_args: Vec<&ast::Type> = match trait_type {
-            ast::Type::Generic(g) => g.args.iter().collect(),
-            _ => Vec::new(),
-        };
-        for (i, tp) in trait_decl_type_params
-            .iter()
-            .filter(|p| !p.is_effect)
-            .enumerate()
-        {
-            if self
-                .annotate_ctx
-                .trait_ctx
-                .type_params
-                .contains_key(&tp.name)
-            {
-                continue;
-            }
-            let Some(arg_ast) = trait_args.get(i) else {
-                continue;
-            };
-            let resolved_arg = self.resolve_type(arg_ast);
-            let idx = self.annotate_ctx.trait_ctx.type_params.len() as u32;
-            self.annotate_ctx
-                .trait_ctx
-                .type_params
-                .insert(tp.name.clone(), (idx, resolved_arg));
-            if !tp.bounds.is_empty() {
-                self.annotate_ctx
-                    .trait_ctx
-                    .type_param_bounds
-                    .entry(tp.name.clone())
-                    .or_default()
-                    .extend(tp.bounds.clone());
-            }
-        }
-    }
 }
 
 /// Extract a type name from an AST type without needing an Elaborator instance.
